requirements.txt
fastapi
uvicorn
openai
pydantic
python-dotenv
langchain
llama-index
pandas